Adhere to Forge naming standards
