+++ src/main/java/net/minecraftforge/registries/ObjectHolderRef.java
@@ -19,56 +19,125 @@
 
 package net.minecraftforge.registries;
 
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.Queue;
-import jline.internal.Nullable;
-import net.minecraftforge.fml.common.registry.GameRegistry;
-import net.minecraftforge.fml.common.registry.GameRegistry.ObjectHolder;
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import javax.annotation.Nullable;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.fml.common.FMLLog;
 
 /**
- * Used in tracking {@link ObjectHolder} references.
+ * Internal default implementation for {@link IObjectHolderRef}.
  */
-interface ObjectHolderRef
+class ObjectHolderRef implements IObjectHolderRef
 {
-    /**
-     * Updates the referenced object holder.
-     * Could be setting a field or invalidating some cache.
-     */
-    void apply();
+    private Field field;
+    private ResourceLocation injectedObject;
+    private boolean isValid;
+    private ForgeRegistry<?> registry;
 
-    /**
-     * @return whether the reference is valid (the underlying registry exists etc.)
-     */
-    boolean isValid();
-
-    /**
-     * Helper method for getting the registry associated with a type.
-     *
-     * @param referenceType the type to check
-     * @return the associated registry if the type is indeed registerable, null otherwise
-     */
     @SuppressWarnings("unchecked")
-    @Nullable
-    public static ForgeRegistry<?> getRegistryForType(Class<?> referenceType) {
-        Queue<Class<?>> typesToExamine = new LinkedList<>();
-        typesToExamine.add(referenceType);
+    ObjectHolderRef(Field field, ResourceLocation injectedObject, boolean extractFromExistingValues)
+    {
+        registry = IObjectHolderRef.getRegistryForType(field.getType());
 
-        ForgeRegistry<?> registry = null;
-        while (!typesToExamine.isEmpty() && registry == null)
+        this.field = field;
+        this.isValid = registry != null;
+        if (extractFromExistingValues)
         {
-            Class<?> type = typesToExamine.remove();
-            Collections.addAll(typesToExamine, type.getInterfaces());
-            if (IForgeRegistryEntry.class.isAssignableFrom(type))
+            try
             {
-                registry = (ForgeRegistry<?>) GameRegistry.findRegistry((Class<IForgeRegistryEntry>) type);
-                final Class<?> parentType = type.getSuperclass();
-                if (parentType != null)
+                Object existing = field.get(null);
+                // nothing is ever allowed to replace AIR
+                if (existing == null || existing == registry.getDefault())
                 {
-                    typesToExamine.add(parentType);
+                    this.injectedObject = null;
+                    this.field = null;
+                    this.isValid = false;
+                    return;
+                } else
+                {
+                    this.injectedObject = ((IForgeRegistryEntry) existing).getRegistryName();
                 }
+            } catch (IllegalAccessException e)
+            {
+                throw new RuntimeException(e);
             }
+        } else
+        {
+            this.injectedObject = injectedObject;
         }
-        return registry;
+
+        if (this.injectedObject == null || !isValid())
+        {
+            throw new IllegalStateException(String.format("The ObjectHolder annotation cannot apply to a field that does not map to a registry. Ensure the registry was created during the RegistryEvent.NewRegistry event. (found : %s at %s.%s)", field.getType().getName(), field.getClass().getName(), field.getName()));
+        }
+        try
+        {
+            FinalFieldHelper.makeWritable(field);
+        } catch (ReflectiveOperationException e)
+        {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public boolean isValid()
+    {
+        return isValid;
+    }
+
+    public void apply()
+    {
+        Object thing;
+        if (isValid && registry.containsKey(injectedObject) && !registry.isDummied(injectedObject))
+        {
+            thing = registry.getValue(injectedObject);
+        } else
+        {
+            thing = null;
+        }
+
+        if (thing == null)
+        {
+            FMLLog.log.debug("Unable to lookup {} for {}. This means the object wasn't registered. It's likely just mod options.", injectedObject, field);
+            return;
+        }
+        try
+        {
+            FinalFieldHelper.setField(field, null, thing);
+        } catch (IllegalArgumentException | ReflectiveOperationException e)
+        {
+            FMLLog.log.warn("Unable to set {} with value {} ({})", this.field, thing, this.injectedObject, e);
+        }
+    }
+
+    private static class FinalFieldHelper
+    {
+        private static Field modifiersField;
+        private static Object reflectionFactory;
+        private static Method newFieldAccessor;
+        private static Method fieldAccessorSet;
+
+        static Field makeWritable(Field f) throws ReflectiveOperationException
+        {
+            f.setAccessible(true);
+            if (modifiersField == null)
+            {
+                Method getReflectionFactory = Class.forName("sun.reflect.ReflectionFactory").getDeclaredMethod("getReflectionFactory");
+                reflectionFactory = getReflectionFactory.invoke(null);
+                newFieldAccessor = Class.forName("sun.reflect.ReflectionFactory").getDeclaredMethod("newFieldAccessor", Field.class, boolean.class);
+                fieldAccessorSet = Class.forName("sun.reflect.FieldAccessor").getDeclaredMethod("set", Object.class, Object.class);
+                modifiersField = Field.class.getDeclaredField("modifiers");
+                modifiersField.setAccessible(true);
+            }
+            modifiersField.setInt(f, f.getModifiers() & ~Modifier.FINAL);
+            return f;
+        }
+
+        static void setField(Field field, @Nullable Object instance, Object thing) throws ReflectiveOperationException
+        {
+            Object fieldAccessor = newFieldAccessor.invoke(reflectionFactory, field, false);
+            fieldAccessorSet.invoke(fieldAccessor, instance, thing);
+        }
     }
 }
